% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_siane_map.R
\name{get_siane_map}
\alias{get_siane_map}
\title{get_siane_map}
\usage{
get_siane_map(obj, level, year, canarias)
}
\arguments{
\item{obj}{: Path of the Siane maps.}

\item{level}{: Is the administrative level. Can take the values of "Provincia","Municipio" or "Comunidades"}

\item{year}{: The numeric year of the maps requested.}

\item{canarias}{: Canarias set to TRUE returns the Canarias map. Canarias set to FALSE returns to the Peninsulae maps.}
}
\value{
A S4 object specified by the user
}
\description{
A function that returns a S4 object of a specific map of Spain
}
