% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_siane.R
\name{plot_siane}
\alias{plot_siane}
\title{plot_siane}
\usage{
plot_siane(shp, ine_path, subsetvars, pallete_colour, n, style)
}
\arguments{
\item{shp}{: Is the shapefile return by the /code{get_siane_map} function.}

\item{subsetvars}{: A vector of expressions. These expresions have to define univocally the values that we are plotting.}

\item{pallete_colour}{: A pallete_colour of the RColorBrewer package. It is a parameter from the /code{RColorBrewer::brewer.pal} function}

\item{n}{: The number of breaks in the pallete. It is a parameter from the /code{classInt::classIntervals} function}

\item{style}{: The way the breaks are numerically distributed. It is a parameter from the /code{classInt::classIntervals} function}

\item{inepath}{: The path of the Ine Data we want to plot.}
}
\description{
This is a function that plots coloured polygons of the territories over the map.
}
