% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siane_map.R
\name{siane_map}
\alias{siane_map}
\title{Load a spanish map}
\usage{
siane_map(obj, canarias, year, level, scale)
}
\arguments{
\item{obj}{: Path to the Siane maps.}

\item{canarias}{: Canarias set to TRUE returns the Canarias map. Canarias set to FALSE returns to the Peninsulae maps.}

\item{year}{: The numeric year of the maps requested.}

\item{level}{: Is the administrative level. Can take the values of "Provincia","Municipio" or "Comunidades"}

\item{scale}{: It's the scale of the map.}
}
\value{
A S4 object specified by the user
}
\description{
A function that returns a S4 object of a Spain specific map
}
