% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siane_merge.R
\name{siane_merge}
\alias{siane_merge}
\title{Merging data and maps}
\usage{
siane_merge(shp, df, by, level, value)
}
\arguments{
\item{shp}{: A S4 object returned by the /code{get_siane_map} function.}

\item{df}{: The data frame with the data we want to plot}

\item{by}{: The name of the column that contains the ID's of the territories.}

\item{level}{: The level of the territory.}

\item{value}{: The statistical values's we want to plot. Only one value per territory.}
}
\value{
A S4 object with the statistical data provided.
}
\description{
This is a function that binds data to polygons
}
\details{
- \code{shp} is a S4 object. \cr
\code{df} is a data frame. This data frame should have at least two columns: one column of statistical data and another column of territorial codes. 
There will be as many missing territories as missing codes in the dataframe. \cr
\code{by} is the dataframe's codes column. \cr 
For \code{level} parameter choose one: "Municipios", "Provincias", "Comunidades" . It's very important that this level is consistent with the S4 object. \cr
\code{value} is the name of the column with the statistical values
}
