% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siane_plot.R
\name{siane_plot}
\alias{siane_plot}
\title{Plot a map}
\usage{
siane_plot(shp, pallete_colour, n, style, ...)
}
\arguments{
\item{shp}{: A S4 object returned by the /code{siane_merge} function.}

\item{pallete_colour}{: A pallete_colour of the RColorBrewer package.}

\item{n}{: The number of breaks in the pallete.}

\item{style}{: The way the breaks are numerically distributed.}
}
\description{
This is a function that plots coloured polygons of the territories over the map.
}
\details{
- \code{shp} is a S4 object. \cr
\code{pallete_colour} is a parameter from the /code{RColorBrewer::brewer.pal} function. The default pallete is "OrRd". Check RColorBrewer:display.brewer.all() to list all possible colours.
\code{n} is a parameter from the /code{classInt::classIntervals} function. Default  \code{n = 5}
\code{style} is a parameter from the /code{classInt::classIntervals} function. Defauly \code{style = quantile} Check ?classInt::classIntervals to list more options.
\code{...} other parameters can be set to the legend() function and the title() function.
Please check ?legend and ?title to learn about the parameters of these functions.
}
