% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_subset.R
\name{suggest_subset}
\alias{suggest_subset}
\title{suggest_subset}
\usage{
suggest_subset(path_ine)
}
\arguments{
\item{path_ine}{: Is the path of the INE pc-axis file.}
}
\description{
This function allows the user to create a random subset of vars to plot
}
