% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_siane.R
\name{plot_siane}
\alias{plot_siane}
\title{plot_siane}
\usage{
plot_siane(shp, df, by, level, value, pallete_colour, n, style, title, x, ...)
}
\arguments{
\item{shp}{: Is the shapefile return by the /code{get_siane_map} function.}

\item{df}{: The name of the data frame with the data we want to plot}

\item{by}{: The name of the column that contains the ID's of the territories.}

\item{level}{: The level of the territory: "Municipios", "Provincias", "Comunidades"}

\item{value}{: A vector of expressions. These expresions have to define univocally the values that we are plotting.}

\item{pallete_colour}{: A pallete_colour of the RColorBrewer package. It is a parameter from the /code{RColorBrewer::brewer.pal} function}

\item{n}{: The number of breaks in the pallete. It is a parameter from the /code{classInt::classIntervals} function}

\item{style}{: The way the breaks are numerically distributed. It is a parameter from the /code{classInt::classIntervals} function}

\item{title}{: It's the title of the plot}
}
\description{
This is a function that plots coloured polygons of the territories over the map.
}
