% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_siane.R
\name{register_siane}
\alias{register_siane}
\title{register_siane}
\usage{
register_siane(path)
}
\arguments{
\item{path}{: The path of the maps at it's highest hierarchical level}
}
\description{
A function that helps loading the path of Siane maps
}
