% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siane_merge.R
\name{siane_merge}
\alias{siane_merge}
\title{Merges data and maps}
\usage{
siane_merge(shp, df, by, level, value)
}
\arguments{
\item{shp}{: A S4 object returned by the \code{siane_map} function.}

\item{df}{: The data frame with the data we want to plot}

\item{by}{: The name of the column that contains the ID's of the territories.}

\item{level}{: The level of the territory.}

\item{value}{: The statistical values's we want to plot. Only one value per territory.}
}
\value{
A S4 object with the statistical data provided.
}
\description{
This is a function that binds data to polygons
}
\details{
- \code{shp} is a S4 object returned by the \code{siane_map} function. \cr
- \code{df} is a data frame. This data frame should have at least two columns:
 one column of statistical data and another column of territorial codes. 
There will be as many missing territories as missing codes in the dataframe. \cr
- \code{by} is the dataframe's codes column. These codes should be the
 INE's official codes. \cr 
- For the \code{level} parameter you pick one of these: "Municipios",
 "Provincias", "Comunidades" . It's very important that this level
  is consistent with the S4 object level. \cr
- \code{value} is the name of the column with the statistical values
}
\examples{
\dontrun{
library(pxR)
obj <- register_siane("/home/ncarvalho/Downloads/")
shp <- siane_map(obj = obj, level = "Municipios", canarias = FALSE) # Loading the municipality's map of Spain
plot(shp) # Plot the map
df <- as.data.frame(read.px("/home/ncarvalho/Downloads/2879.px"))# Get this file in INE's website. Check the README to learn how to do it.
by <- "codes"
df[[by]] <- sapply(df$Municipios,function(x) strsplit(x = as.character(x), split = " ")[[1]][1])
year <- 2016
df$Periodo <- as.numeric(as.character(df$Periodo))
df <- df[df$Sex == "Total" & df$Periodo == year, ]
value <- "value"
level <- "Municipios"
shp_merged <- siane_merge(shp = shp, df = df,by = by, level = level, value = value)
plot(shp_merged)}
 
}
