% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siane_plot.R
\name{siane_plot}
\alias{siane_plot}
\title{Plot a map}
\usage{
siane_plot(shp, pallete_colour, n, style, ...)
}
\arguments{
\item{shp}{: A S4 object returned by the /code{siane_merge} function.}

\item{pallete_colour}{: A pallete_colour of the RColorBrewer package.}

\item{n}{: The number of breaks in the pallete.}

\item{style}{: The way the breaks are numerically distributed.}
}
\description{
This is a function that plots coloured polygons of the territories over the map.
}
\details{
- \code{shp} is a S4 object returned by the \code{siane_merge} function . \cr
\code{pallete_colour} is a parameter from the \code{RColorBrewer::brewer.pal} function. The default pallete is "OrRd". Check RColorBrewer:display.brewer.all() to list all possible colours.
\code{n} is a parameter from the /code{classInt::classIntervals} function. Default  \code{n = 5}  \cr
\code{style} is a parameter from the /code{classInt::classIntervals} function. Defauly \code{style = quantile} Check ?classInt::classIntervals to list more options. \cr
Please check ?legend and ?title to learn about the parameters of these functions. \cr
}
\examples{
obj <- register_siane(system.file("extdata",package = "Siane"))
shp <- siane_map(obj = obj, level = "Municipios", canarias = FALSE) # Loading the municipality's map of Spain
plot(shp) # Plot the map
df <- read.csv(system.file("extdata","sampleRiojaPopulation.csv",package = "Siane"))
by <- "codes"
df[[by]] <- sapply(df$Municipios,function(x) strsplit(x = as.character(x), split = " ")[[1]][1])
year <- 2016
df$Periodo <- as.numeric(as.character(df$Periodo))
df <- df[df$Sex == "Total" & df$Periodo == year, ]
value <- "value"
level <- "Municipios"
shp_merged <- siane_merge(shp = shp, df = df,by = by, level = level, value = value)
plot(shp_merged)
siane_plot(shp = shp_merged, x = "bottomright", main = "Population of La Rioja")
 
 
 
}
